chore(): clean debug page
