+++ package.json
@@ -26,6 +26,7 @@
     "@radix-ui/react-tabs": "^1.1.13",
     "@tanstack/react-query": "^5.59.20",
     "@use-gesture/react": "^10.3.1",
+    "@yudiel/react-qr-scanner": "^2.5.0",
     "axios": "^1.13.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
